Removed OnActivate as an event. It caused problems when adding components after an entity had been created. Now we only use OnInit callbacks, since it's basically equally fast as doing the other approach. Fixed issues with models being added to visibility during vis update. Started working on replacing the game entity with a new copy when updating the editor entity.

+++ code/application/game/dataset.h
@@ -38,7 +38,7 @@ struct Dataset
         /// partition identifier
         uint16_t partitionId = 0xFFFF;
         /// number of instances in view
-        uint32_t numInstances = 0;
+        uint16_t numInstances = 0;
         /// component buffers. @note Can be NULL if a queried component has no fields
         void* buffers[MAX_COMPONENT_BUFFERS];
         /// which instances are valid in this buffer
